fixed line ends in PM history quick search position saved fixed that 3 standard OP commands were missing in some cases menu/toolbar/quicksearch position and size are saved updated PCRE library stopping old search request when starting new one rewritten search queue + search interval is hub dependent (idea from RevConnect) added missing searchresult's context menu header

+++ windows/HubFrame.cpp
@@ -958,7 +958,7 @@ void HubFrame::findText(tstring const& needle) throw() {
 	// found? set selection
 	if(currentNeedlePos != -1) {
 		ft.chrg.cpMin = currentNeedlePos;
-		ft.chrg.cpMax = currentNeedlePos + needle.length();
+		ft.chrg.cpMax = currentNeedlePos + (long)needle.length();
 		ctrlClient.SetFocus();
 		ctrlClient.SendMessage(EM_EXSETSEL, 0, (LPARAM)&ft);
 	} else {
